backend/requirements.txt
cd backend
echo "numpy
pandas
google-generativeai
yfinance
requests
torch
scikit-learn
matplotlib
fastapi
uvicorn" > requirements.txt

pip install -r requirements.txt
